Moved types to CANUtils, fixed bug
