+++ src/CAN/CANUtils.h
@@ -1,14 +1,30 @@
 #pragma once
 
-#include "CAN.h"
-
 extern "C" {
 #include "../HindsightCAN/CANPacket.h"
 }
 
 #include <chrono>
+#include <utility>
 
 namespace can {
+enum class devicegroup_t {
+	broadcast,
+	reserved,
+	master,
+	power,
+	motor,
+	telemetry,
+	gpio,
+	science
+};
+
+using deviceserial_t = uint8_t;
+using telemtype_t = uint8_t; // TODO: convert to enum
+using telemetry_t = int32_t;
+
+using deviceid_t = std::pair<devicegroup_t, deviceserial_t>;
+
 devicegroup_t getDeviceGroup(const CANPacket& packet);
 
 devicegroup_t getDeviceGroup(uint8_t groupCode);
